Introduce gtvj\Page and PSR-4 autoloading
The Page class is introduced to provide data to Slim (for demo
purposes). Page is within a 'gtvj' namespace.

A Page object is then instantiated in index.php to provide data to
the $response object body.

The gtvj namespace is added to the "autoload" within composer.json
so that Page can be autoloaded in the same way as Composer
dependencies.